+++ Libraries/Ultrasound/Ultrasound.h
@@ -16,7 +16,7 @@
 class Ultrasound {
 	public:
 		//Constructors
-		Ultrasound(byte usTrigger, byte usEcho, bool simFlag, int maxRange);
+		Ultrasound(byte usTrigger, byte usEcho, bool simFlag, float maxRange);
 		
 		//Functions
 		bool collisionDetection(float boundry);
@@ -30,7 +30,7 @@ class Ultrasound {
 		//Variables
 		byte _usTrigger, _usEcho;
 		bool _simFlag;
-		int _maxRange;
+		float _maxRange;
 };
 
 #endif
